# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchAlertRequest": {
        "description": "Request body for batch alert operations.",
        "properties": {
          "alert_ids": {
            "items": {
              "type": "string"
            },
            "title": "Alert Ids",
            "type": "array"
          }
        },
        "required": [
          "alert_ids"
        ],
        "title": "BatchAlertRequest",
        "type": "object"
      },
...
      "FrontendAlert": {
        "additionalProperties": true,
        "properties": {
          "bed": {
            "type": "string"
          },
          "closedBy": {
            "type": [
              "string",
              "null"
            ]
          },
          "closureOrigin": {
            "enum": [
              "equipe",
              "sensor",
              "sistema",
              null
            ],
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "lastRepositioning": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "nextRepositioning": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "patientName": {
            "type": "string"
          },
          "riskLevel": {
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "type": "string"
          },
          "room": {
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "acknowledged",
              "completed"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "patientName",
          "riskLevel",
          "status"
        ],
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Monitor de Alertas UPP",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/alerts/export/csv": {
      "get": {
        "description": "Exporta alertas em formato CSV.\n\nQuery Parameters:\n- start_date: Data inicial no formato YYYY-MM-DD\n- end_date: Data final no formato YYYY-MM-DD\n- status: Status do alerta (pending, acknowledged, completed)\n- patient_id: ID do paciente\n- limit: Limite de registros (máximo 100000)\n\nReturns:\n- CSV file with alerts data\n\nAutenticação exigida via get_current_user (JWT em Bearer ou cookie\naccess_token). Exporta dados clínicos de pacientes — nunca deixar sem auth.",
        "operationId": "export_alerts_csv_api_alerts_export_csv_get",
        "parameters": [
          {
            "description": "Data inicial (YYYY-MM-DD)",
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Data inicial (YYYY-MM-DD)",
              "title": "Start Date"
            }
          },
          {
            "description": "Data final (YYYY-MM-DD)",
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Data final (YYYY-MM-DD)",
              "title": "End Date"
            }
          },
          {
            "description": "Status: pending, acknowledged, completed",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Status: pending, acknowledged, completed",
              "title": "Status"
            }
          },
          {
            "description": "ID do paciente",
            "in": "query",
            "name": "patient_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ID do paciente",
              "title": "Patient Id"
            }
          },
          {
            "description": "Limite de registros",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10000,
              "description": "Limite de registros",
              "maximum": 100000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Alerts Csv",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/alerts/export/pdf": {
      "get": {
        "description": "Exporta alertas em formato PDF.\n\nQuery Parameters:\n- start_date: Data inicial no formato YYYY-MM-DD\n- end_date: Data final no formato YYYY-MM-DD\n- status: Status do alerta (pending, acknowledged, completed)\n- patient_id: ID do paciente\n\nReturns:\n- PDF file with formatted alerts report\n\nAutenticação exigida via get_current_user (JWT em Bearer ou cookie\naccess_token). Exporta dados clínicos de pacientes — nunca deixar sem auth.",
        "operationId": "export_alerts_pdf_api_alerts_export_pdf_get",
        "parameters": [
          {
            "description": "Data inicial (YYYY-MM-DD)",
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Data inicial (YYYY-MM-DD)",
              "title": "Start Date"
            }
          },
          {
            "description": "Data final (YYYY-MM-DD)",
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Data final (YYYY-MM-DD)",
              "title": "End Date"
            }
          },
          {
            "description": "Status: pending, acknowledged, completed",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Status: pending, acknowledged, completed",
              "title": "Status"
            }
          },
          {
            "description": "ID do paciente",
            "in": "query",
            "name": "patient_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ID do paciente",
              "title": "Patient Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Alerts Pdf",
        "tags": [
          "alerts"
        ]
      }
    },
...
    "/api/frontend/alerts": {
      "get": {
        "description": "Return alerts in a shape convenient for the React frontend with optional filters.\n\nQuery Parameters:\n- horas: int (default 24) - Time window in hours\n- riskLevel: 'high'|'medium'|'low' - Filter by risk level\n- status_filter: 'pending'|'acknowledged'|'completed' - Filter by status\n- room: str - Filter by room number (fuzzy match)\n- limit: int (default 100) - Pagination limit\n- offset: int (default 0) - Pagination offset\n\nEach alert contains: id, patientName, room, bed, lastRepositioning (ISO),\nnextRepositioning (ISO), riskLevel (high|medium|low), status (pending|acknowledged|completed)\n\nCabecalho `X-Total-Count`: quantos alertas casam com os filtros, ANTES do\ncorte por `limit`. O corpo continua sendo o array puro (o contrato que o\nfrontend e os testes usam), mas sem esse numero a resposta omite em\nsilencio: uma lista de 100 com `limit=100` e indistinguivel de \"existem\nexatamente 100\", e o dashboard filtra em memoria sobre o que recebeu — um\npaciente atrasado ficaria fora da tela sem nenhum sinal.",
        "operationId": "frontend_alerts_api_frontend_alerts_get",
        "parameters": [
          {
            "in": "query",
            "name": "horas",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "default": 24,
              "title": "Horas"
            }
          },
          {
            "in": "query",
            "name": "riskLevel",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risklevel"
            }
          },
          {
            "in": "query",
            "name": "status_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "in": "query",
            "name": "room",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Room"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FrontendAlert"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Frontend Alerts",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/frontend/alerts/batch/acknowledge": {
      "post": {
        "description": "Acknowledge multiple alerts at once.",
        "operationId": "batch_acknowledge_api_frontend_alerts_batch_acknowledge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchAlertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Batch Acknowledge Api Frontend Alerts Batch Acknowledge Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Batch Acknowledge",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/frontend/alerts/batch/complete": {
      "post": {
        "description": "Complete multiple alerts at once.",
        "operationId": "batch_complete_api_frontend_alerts_batch_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchAlertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Batch Complete Api Frontend Alerts Batch Complete Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Batch Complete",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/frontend/alerts/{alert_id}/acknowledge": {
      "post": {
        "description": "Reconhece um alerta e registra evento na timeline.",
        "operationId": "frontend_acknowledge_api_frontend_alerts__alert_id__acknowledge_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Frontend Acknowledge Api Frontend Alerts  Alert Id  Acknowledge Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Frontend Acknowledge",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/frontend/alerts/{alert_id}/complete": {
      "post": {
        "description": "Completa/fecha um alerta e registra evento na timeline.",
        "operationId": "frontend_complete_api_frontend_alerts__alert_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Frontend Complete Api Frontend Alerts  Alert Id  Complete Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Frontend Complete",
        "tags": [
          "alerts"
        ]
      }
    },
...
  }
}